compiler: split ItemProcessors, FileProcessors and Caching
